pyproject.toml
[project]
name = "sellio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.2"
dependencies = [
    "alembic>=1.15.1",
    "argon2-cffi==23.1.0",
    "asyncpg==0.30.0",
    "fastapi[standard]==0.115.11",
    "graphql-core-next==1.1.1",
    "hiku==0.8.0rc8",
    "prometheus-client==0.21.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "sqlalchemy[asyncio]==2.0.39",
    "toml==0.10.2",
]

[dependency-groups]
dev = [
    "ruff==0.11.1",
]
